SVM Kernal trick worked
As the data set is having random classifications and we don't have enough proofs to justify the results, we have no much choice other than playing around with the pre-processing , data transformations and hyperparameter tunings.

as SVM was the model, which gave a little better Recall value, I have desided to play around with the kernal tricks. 
As the "linear" kernal would always try to separate the data by drawing a linear plane between the data to separate the 2D future space.
Since the data is not linear and it requires a polygonal separation , I have changed the kernel to polynomial and it did the magic. We have achived the maximum of 100% Recall.